requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Elasticsearch
elasticsearch==8.11.0

# Google Cloud
google-cloud-storage==2.10.0
google-auth==2.23.4

# LangChain
langchain==0.0.350
langchain-community==0.0.5

# Document processing
pypdf==3.17.4
pdfplumber==0.10.3
python-multipart==0.0.6
Pillow==10.1.0
pytesseract==0.3.10
pandas==2.1.4
openpyxl==3.1.2

# Async support
aiofiles==23.2.1
httpx==0.25.2

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
coverage==7.3.2

# Configuration
PyYAML==6.0.1
python-dotenv==1.0.0
